Work around the exception raised when a base path doesn't exist
This fix is not 100% guaranteed to be proper, but it solved the issue at
the moment. It should be confirmed whether such approach is acceptable,
and if not, what should be done. FSDBManager.get_root_dir raises an
exception when the folder it tries to access doesn't exist, which is a
possibility in case of chroots.