Handle deleted files in refresh_path method
The update_db family of functions are unable to detect missing
filesystem objects among it's existing entries since they basically
reindex from the filesystem, keeping all existing data. This patch adds
a branch to the refresh_path method to explicitly check for the
existence of the given path, and in case it's missing, call the remove
function, otherwise perform normal reindexing as usual.